Added seekbars for rudder and throttle, need to find solution for connectionstatus. Next joystick.

+++ app/src/main/java/com/example/javaflightgearandroidapp/Model/ActiveClientModel.java
@@ -43,7 +43,7 @@ public class ActiveClientModel {
         } catch(Exception e) { }
     }
 
-    public void write(String command, int value) {
+    public void write(String command, float value) {
         try {
             dispatchQueue.put(new Runnable() {
                 @Override
